Create a opaque newtype for `RawMiniTable` and `RawMiniTableEnum`
These types have to be exposed to gencode so that it can implement the
`AssociatedMiniTable` and `AssociatedMiniTableEnum` traits. Wrapping the
pointer in an opaque type is useful, though, to expose as little as possible of
implementation details.

I thought about creating a separate type that wraps `RawMiniTable` instead of
just wrapping the `NonNull`, but I think that adds an unnecessary layer of
abstraction. For now the type exposed to the gencode is the same type used for
FFI, but we can always change that later if necessary.

PiperOrigin-RevId: 806031508